Crate incidents

Incidents is a library for Rust that explores error handling. The idea is to eventually get an RFC together with the ideas and concepts of the library.

The Error Trait

Errors are arbitrary, they just need to implement the Error trait. This trait currently differs from the one in the std::error module and has different associated functions.

The bare minimum an error needs to implement is name which should be the capitalized name of the error as it should be displayed to a user if it appears in error messages.

Example:

struct BadOperation {
    desc: &str,
}

impl Error for BadOperation {
    fn name(&self) -> &str { "Bad Operation" }
    fn description(&self) -> Option<&str> { Some(self.desc) }
}

Failures

Incidents provides a special wrapper type called a Failure<T> which will automatically box an error. In addition to moving the error to the heap, which will keep results small, this also adds automatic debug helpers for debug builds. In release the failure is represented as a simple box, whereas in debug it actually builds a Traceback behind the scenes that aids debugging.

A failure can be dereferenced which will return the error.

Throwing Errors

Errors can be thrown with the fail! macro and be automatically propagated with the try! macro. For this to work the return value needs to be a Result with a Failure<E: Error> or Error on the error side. For the latter a FResult is provided.

Example:

fn function_that_fails(good: bool) -> FResult<int, TheError> {
    if !good {
        fail!(TheError { desc: "something went wrong" });
    }
    Ok(23)
}

fn function_that_tries() -> FResult<int, TheError> {
    let rv = try!(function_that_fails());
    Ok(rv + 42)
}

Handling Errors

Errors that are in failures can be handled very transparently. The failure is strongly typed and derefs into the actual error:

fn function_that_handles() {
    match function_that_fails() {
        Ok(result) => {
            println!("The result: {}", result);
        }
        Err(ref err) => {
            println!("Error description: {}", err.desc);
        }
    }
}

Error Conversion

Errors can be converted through the use of the FromError trait. This allows a library to easily work with other libraries that it might use internally and give the user a consistent error experience.

Example:

enum MyError {
    BadOperation,
    IoError(io::IoError),
}

impl Error for MyError {
    fn name(&self) -> &str {
        match *self {
            MyError::BadOperation => "Bad Operation",
            MyError::IoError(ref err) => err.name(),
        }
    }
}

impl FromError<IoError> for MyError {
    fn from_error(err: io::IoError) -> MyError {
        MyError::IoError(err)
    }
}

Tracebacks

In debug builds tracebacks are available. At any point can you call Traceback::from_failure() with a failure to get the traceback. This also works in release builds but the return value will be None.

Tracebacks can also be printed with print_traceback and TraceFormatter. This looks a bit like Python tracebacks:

Traceback (most recent cause last):
  File "/Users/mitsuhiko/Development/rust-incidents/examples/basic.rs", line 37
    fail!(FileNotFound { file: Some(Path::new("/missing.txt")) });
  File "/Users/mitsuhiko/Development/rust-incidents/examples/basic.rs", line 41
    try!(testing());
File Not Found: the file does not exist (file=/missing.txt)

Macros

fail!
Fails with an error.

try!
Tries to unwrap the result or propagates an error.

Structs

Failure
A failure wraps an error in a box.

LocationInfo
This struct provides error location information.

TraceFormatter
Formats tracebacks.

Traceback
Encapsulates errors.

Traits

ConstructFailure
A trait that is used to construct failures.

Error
Represents an error.

ErrorExt
Adds dynamic helpers to errors.

Frame
Frames are an internal construct in incidents for debugging.

FromError
A trait for types that can be converted from a given error type E.

Functions

print_traceback
Print the traceback of a failure.

Type Definitions

FResult
A result type which wraps the error in a failure.