Function incidents::print_traceback

pub fn print_traceback<E: Error>(failure: &Failure<E>)

Print the traceback of a failure.

In release builds this will not contain an actual traceback but an emulation based on the information that is available on the error itself.

This internally uses TraceFormatter.