Struct incidents::Failure

pub struct Failure<E: Error> {
    // some fields omitted
}

A failure wraps an error in a box.

This solves two purposes. The first one is that it moves the error to the heap which causes the result value to become smaller for the optimal (non error) case. The second purpose is to hold additional information for debug builds that aid debugging.

In debug builds a failure holds a Traceback so that it can become possible to discover how an error came about.

Failures can be dereferenced to get the error that created them.

To get the traceback of a failure use the Traceback::from_failure function.