Struct incidents::LocationInfo

pub struct LocationInfo {
    // some fields omitted
}

This struct provides error location information.

This is intentionally kept private internally so that at a later point it can be changed to not actually contain the debug information but to only hold a frozen PC value and later read the debug info from DWARF.

Methods

impl LocationInfo

fn new(file: &str, line: uint) -> LocationInfo

Creates a new location info from specific values.

fn get_source_line(&self) -> IoResult<String>

Returns the source line which caused the error.

This loads the file from the file system and fetches the line mentioned in the location info. This might be quite an expensive operation and it can also return invalid data in case the file was changed after the compilation took place.

fn file(&self) -> &Path

Get the filename of this location info.

fn line(&self) -> uint

Get the line number of this location info.