Struct incidents::TraceFormatter

pub struct TraceFormatter<W: Writer> {
    // some fields omitted
}

Formats tracebacks.

Methods

impl<W: Writer> TraceFormatter<W>

The trace formatter can be used to render tracebacks.

fn new(writer: W) -> TraceFormatter<W>

Constructs a new trace formatter.

fn format_fallback_trace(&mut self, err: &Error) -> IoResult<()>

Formats a fallback trace from an error.

This is useful for non debug builds where tracebacks are genereally not available. In that case you can still get some debug information from the errors alone. This will try to find as much information as possible and print them out in a similar format.

fn format_traces(&mut self, i: &Traceback) -> IoResult<()>

Formats all traces of the incident.

fn format_frame(&mut self, frm: &Frame) -> IoResult<()>

Formats a single frame.

fn format_cause(&mut self, err: &Error) -> IoResult<()>

Formats the cause of an error.