Struct incidents::Traceback

pub struct Traceback {
    // some fields omitted
}

Encapsulates errors.

Tracebacks are generally only available in debug builds. In debug builds they associate errors with as much debug information as possible.

Methods

impl Traceback

fn from_failure<E: Error>(failure: &Failure<E>) -> Option<&Traceback>

Return the traceback object from a failure.

This will return a borrowed reference to the traceback contained in a failure. This operation will return None for release builds for which a failure does not actually contain a traceback.

fn frame(&self) -> &Frame + Send

Returns the first frame of the traceback.

fn error(&self) -> &Error

Returns the error of the traceback.

Note that a traceback can be chained to another error behind the scenes. This will only ever give you one error and it will always be the most recent one. To discover causes you will need to traverse the frames manually.

Example usage:

println!("Error: {}", traceback.error().name());

fn error_frame(&self) -> &Frame + Send

Returns the frame that contains the error.

fn error_as<E: Error>(&self) -> Option<&E>

Returns the error as a pointer to a concrete type.

Example usage:

match traceback.error_as::<ResourceNotFound>() {
    Some(rnf) => {
        println!("Resource not found: {}", rnf.resource);
    },
    None => {}
}

fn is<E: Error>(&self) -> bool

Checks if the error is of a specific type.

Example usage:

if traceback.is::<ResourceNotFound>() {
    println!("Looks like a resource is missing.");
}

fn name(&self) -> &str

Shorthand to get the name of the error.

fn description(&self) -> Option<&str>

Gets the description the error.

fn detail(&self) -> Option<String>

Gets the detail of the error.

fn traceback(&self) -> Vec<(&Error, Vec<&Frame + Send>)>

Returns the traces in this traceback.

Keep in mind that it is entirely permissible for the return value to be an empty vector. This can for instance happen in release builds.