Trait incidents::ConstructFailure

pub trait ConstructFailure<A> {
    fn construct_failure(args: A, loc: Option<LocationInfo>) -> Self;
}

A trait that is used to construct failures.

Generally this trait should currently be considered internal as there are no public ways to construct Failure objects. This is used by fail! and try! to fail and convert errors.

Required Methods

fn construct_failure(args: A, loc: Option<LocationInfo>) -> Self