Trait incidents::Frame

pub trait Frame: Send {
    fn error(&self) -> Option<&Error> { ... }
    fn location(&self) -> Option<&LocationInfo> { ... }
    fn cause_frame(&self) -> Option<&Frame + Send> { ... }
    fn previous_frame(&self) -> Option<&Frame + Send> { ... }
    fn trace(&self) -> Option<(&Error, Vec<&Frame + Send>)> { ... }
}

Frames are an internal construct in incidents for debugging.

Frames primarily are used to chain errors together or to attach additional debug information to them. For the most part you can ignore that frames exist. The primary use for frames is to render stacktraces.

Provided Methods

fn error(&self) -> Option<&Error>

If the frame points to an error, this returns it.

fn location(&self) -> Option<&LocationInfo>

If the frame contains location information, this returns it.

fn cause_frame(&self) -> Option<&Frame + Send>

If the frame was caused by another frame, this returns a reference to it. This is used if an error gets handled by failing with another error.

fn previous_frame(&self) -> Option<&Frame + Send>

If the frame is linked directly to another frame, this returns a reference to it. This will never be the cause frame.

fn trace(&self) -> Option<(&Error, Vec<&Frame + Send>)>

If it is possible to construct a trace from this frame, it will return the error and the trace to it. Note that this is not always possible.